Add context manager to acquire locks with retry and move function to resourcelockrepo
