+++ backend/dataall/modules/shares_base/services/sharing_service.py
@@ -1,9 +1,7 @@
 import logging
-from typing import Any, List, Tuple
 from dataclasses import dataclass
-from time import sleep
 
-from sqlalchemy.orm import Session
+from typing import Any
 from dataall.core.resource_lock.db.resource_lock_repositories import ResourceLockRepository
 from dataall.base.db import Engine
 from dataall.core.environment.db.environment_models import ConsumptionRole, Environment, EnvironmentGroup
@@ -29,9 +27,6 @@ from dataall.modules.shares_base.services.share_exceptions import PrincipalRoleN
 from dataall.base.db.exceptions import ResourceLockTimeout
 
 log = logging.getLogger(__name__)
-
-MAX_RETRIES = 10
-RETRY_INTERVAL = 60
 
 
 @dataclass
@@ -94,44 +89,45 @@ class SharingService:
                         f'Principal role {share_data.share.principalIAMRoleName} is not found.',
                     )
 
-                if not cls.acquire_lock_with_retry(
-                    resources, session, share_data.share.shareUri, share_data.share.__tablename__
+                with ResourceLockRepository.acquire_lock_with_retry(
+                    resources=resources,
+                    session=session,
+                    acquired_by_uri=share_data.share.shareUri,
+                    acquired_by_type=share_data.share.__tablename__,
                 ):
-                    raise ResourceLockTimeout(
-                        'process approved shares',
-                        f'Failed to acquire lock for dataset {share_data.dataset.datasetUri}',
-                    )
-                for type, processor in ShareProcessorManager.SHARING_PROCESSORS.items():
-                    try:
-                        log.info(f'Granting permissions of {type.value}')
-                        shareable_items = ShareObjectRepository.get_share_data_items_by_type(
-                            session,
-                            share_data.share,
-                            processor.shareable_type,
-                            processor.shareable_uri,
-                            status=ShareItemStatus.Share_Approved.value,
-                        )
-                        success = processor.Processor(session, share_data, shareable_items).process_approved_shares()
-                        log.info(f'Sharing {type.value} succeeded = {success}')
-                        if not success:
+                    for type, processor in ShareProcessorManager.SHARING_PROCESSORS.items():
+                        try:
+                            log.info(f'Granting permissions of {type.value}')
+                            shareable_items = ShareObjectRepository.get_share_data_items_by_type(
+                                session,
+                                share_data.share,
+                                processor.shareable_type,
+                                processor.shareable_uri,
+                                status=ShareItemStatus.Share_Approved.value,
+                            )
+                            success = processor.Processor(
+                                session, share_data, shareable_items
+                            ).process_approved_shares()
+                            log.info(f'Sharing {type.value} succeeded = {success}')
+                            if not success:
+                                share_successful = False
+                        except Exception as e:
+                            log.error(f'Error occurred during sharing of {type.value}: {e}')
+                            ShareStatusRepository.update_share_item_status_batch(
+                                session,
+                                share_uri,
+                                old_status=ShareItemStatus.Share_Approved.value,
+                                new_status=ShareItemStatus.Share_Failed.value,
+                                share_item_type=processor.type.value,
+                            )
+                            ShareStatusRepository.update_share_item_status_batch(
+                                session,
+                                share_uri,
+                                old_status=ShareItemStatus.Share_In_Progress.value,
+                                new_status=ShareItemStatus.Share_Failed.value,
+                                share_item_type=processor.type.value,
+                            )
                             share_successful = False
-                    except Exception as e:
-                        log.error(f'Error occurred during sharing of {type.value}: {e}')
-                        ShareStatusRepository.update_share_item_status_batch(
-                            session,
-                            share_uri,
-                            old_status=ShareItemStatus.Share_Approved.value,
-                            new_status=ShareItemStatus.Share_Failed.value,
-                            share_item_type=processor.type.value,
-                        )
-                        ShareStatusRepository.update_share_item_status_batch(
-                            session,
-                            share_uri,
-                            old_status=ShareItemStatus.Share_In_Progress.value,
-                            new_status=ShareItemStatus.Share_Failed.value,
-                            share_item_type=processor.type.value,
-                        )
-                        share_successful = False
                 return share_successful
 
             except Exception as e:
@@ -143,13 +139,6 @@ class SharingService:
             finally:
                 new_share_state = share_object_sm.run_transition(ShareObjectActions.Finish.value)
                 share_object_sm.update_state(session, share_data.share, new_share_state)
-                for resource in resources:
-                    if not ResourceLockRepository.release_lock(
-                        session, resource[0], resource[1], share_data.share.shareUri
-                    ):
-                        log.error(
-                            f'Failed to release lock for resource: resource_uri={resource[0]}, resource_type={resource[1]}'
-                        )
 
     @classmethod
     def revoke_share(cls, engine: Engine, share_uri: str) -> bool:
@@ -200,45 +189,43 @@ class SharingService:
                         f'Principal role {share_data.share.principalIAMRoleName} is not found.',
                     )
 
-                if not cls.acquire_lock_with_retry(
-                    resources, session, share_data.share.shareUri, share_data.share.__tablename__
+                with ResourceLockRepository.acquire_lock_with_retry(
+                    resources=resources,
+                    session=session,
+                    acquired_by_uri=share_data.share.shareUri,
+                    acquired_by_type=share_data.share.__tablename__,
                 ):
-                    raise ResourceLockTimeout(
-                        'process revoked shares',
-                        f'Failed to acquire lock for dataset {share_data.dataset.datasetUri}',
-                    )
-
-                for type, processor in ShareProcessorManager.SHARING_PROCESSORS.items():
-                    try:
-                        shareable_items = ShareObjectRepository.get_share_data_items_by_type(
-                            session,
-                            share_data.share,
-                            processor.shareable_type,
-                            processor.shareable_uri,
-                            status=ShareItemStatus.Revoke_Approved.value,
-                        )
-                        log.info(f'Revoking permissions with {type.value}')
-                        success = processor.Processor(session, share_data, shareable_items).process_revoked_shares()
-                        log.info(f'Revoking {type.value} succeeded = {success}')
-                        if not success:
+                    for type, processor in ShareProcessorManager.SHARING_PROCESSORS.items():
+                        try:
+                            shareable_items = ShareObjectRepository.get_share_data_items_by_type(
+                                session,
+                                share_data.share,
+                                processor.shareable_type,
+                                processor.shareable_uri,
+                                status=ShareItemStatus.Revoke_Approved.value,
+                            )
+                            log.info(f'Revoking permissions with {type.value}')
+                            success = processor.Processor(session, share_data, shareable_items).process_revoked_shares()
+                            log.info(f'Revoking {type.value} succeeded = {success}')
+                            if not success:
+                                revoke_successful = False
+                        except Exception as e:
+                            log.error(f'Error occurred during share revoking of {type.value}: {e}')
+                            ShareStatusRepository.update_share_item_status_batch(
+                                session,
+                                share_uri,
+                                old_status=ShareItemStatus.Revoke_Approved.value,
+                                new_status=ShareItemStatus.Revoke_Failed.value,
+                                share_item_type=processor.type.value,
+                            )
+                            ShareStatusRepository.update_share_item_status_batch(
+                                session,
+                                share_uri,
+                                old_status=ShareItemStatus.Revoke_In_Progress.value,
+                                new_status=ShareItemStatus.Revoke_Failed.value,
+                                share_item_type=processor.type.value,
+                            )
                             revoke_successful = False
-                    except Exception as e:
-                        log.error(f'Error occurred during share revoking of {type.value}: {e}')
-                        ShareStatusRepository.update_share_item_status_batch(
-                            session,
-                            share_uri,
-                            old_status=ShareItemStatus.Revoke_Approved.value,
-                            new_status=ShareItemStatus.Revoke_Failed.value,
-                            share_item_type=processor.type.value,
-                        )
-                        ShareStatusRepository.update_share_item_status_batch(
-                            session,
-                            share_uri,
-                            old_status=ShareItemStatus.Revoke_In_Progress.value,
-                            new_status=ShareItemStatus.Revoke_Failed.value,
-                            share_item_type=processor.type.value,
-                        )
-                        revoke_successful = False
 
                 return revoke_successful
             except Exception as e:
@@ -254,14 +241,6 @@ class SharingService:
                 else:
                     new_share_state = share_sm.run_transition(ShareObjectActions.Finish.value)
                 share_sm.update_state(session, share_data.share, new_share_state)
-
-                for resource in resources:
-                    if not ResourceLockRepository.release_lock(
-                        session, resource[0], resource[1], share_data.share.shareUri
-                    ):
-                        log.error(
-                            f'Failed to release lock for resource: resource_uri={resource[0]}, resource_type={resource[1]}'
-                        )
 
     @classmethod
     def verify_share(
@@ -356,56 +335,46 @@ class SharingService:
                     log.error(f'Failed to get Principal IAM Role {share_data.share.principalIAMRoleName}, exiting...')
                     return False
                 else:
-                    lock_acquired = cls.acquire_lock_with_retry(
-                        resources, session, share_data.share.shareUri, share_data.share.__tablename__
-                    )
-
-                    if not lock_acquired:
-                        log.error(f'Failed to acquire lock for dataset {share_data.dataset.datasetUri}, exiting...')
-                        error_message = f'SHARING PROCESS TIMEOUT: Failed to acquire lock for dataset {share_data.dataset.datasetUri}'
-                        ShareStatusRepository.update_share_item_health_status_batch(
-                            session,
-                            share_uri,
-                            old_status=ShareItemHealthStatus.PendingReApply.value,
-                            new_status=ShareItemHealthStatus.Unhealthy.value,
-                            message=error_message,
-                        )
-                        return False
-
-                    for type, processor in ShareProcessorManager.SHARING_PROCESSORS.items():
-                        try:
-                            log.info(f'Reapplying permissions to {type.value}')
-                            shareable_items = ShareObjectRepository.get_share_data_items_by_type(
-                                session,
-                                share_data.share,
-                                processor.shareable_type,
-                                processor.shareable_uri,
-                                None,
-                                ShareItemHealthStatus.PendingReApply.value,
-                            )
-                            success = processor.Processor(
-                                session, share_data, shareable_items
-                            ).process_approved_shares()
-                            log.info(f'Reapplying {type.value} succeeded = {success}')
-                            if not success:
-                                reapply_successful = False
-                        except Exception as e:
-                            log.error(f'Error occurred during share reapplying of {type.value}: {e}')
+                    with ResourceLockRepository.acquire_lock_with_retry(
+                        resources=resources,
+                        session=session,
+                        acquired_by_uri=share_data.share.shareUri,
+                        acquired_by_type=share_data.share.__tablename__,
+                    ):
+                        for type, processor in ShareProcessorManager.SHARING_PROCESSORS.items():
+                            try:
+                                log.info(f'Reapplying permissions to {type.value}')
+                                shareable_items = ShareObjectRepository.get_share_data_items_by_type(
+                                    session,
+                                    share_data.share,
+                                    processor.shareable_type,
+                                    processor.shareable_uri,
+                                    None,
+                                    ShareItemHealthStatus.PendingReApply.value,
+                                )
+                                success = processor.Processor(
+                                    session, share_data, shareable_items
+                                ).process_approved_shares()
+                                log.info(f'Reapplying {type.value} succeeded = {success}')
+                                if not success:
+                                    reapply_successful = False
+                            except Exception as e:
+                                log.error(f'Error occurred during share reapplying of {type.value}: {e}')
 
                 return reapply_successful
+
+            except ResourceLockTimeout as e:
+                ShareStatusRepository.update_share_item_health_status_batch(
+                    session,
+                    share_uri,
+                    old_status=ShareItemHealthStatus.PendingReApply.value,
+                    new_status=ShareItemHealthStatus.Unhealthy.value,
+                    message=str(e),
+                )
+
             except Exception as e:
                 log.error(f'Error occurred during share approval: {e}')
                 return False
-
-            finally:
-                with engine.scoped_session() as session:
-                    for resource in resources:
-                        if not ResourceLockRepository.release_lock(
-                            session, resource[0], resource[1], share_data.share.shareUri
-                        ):
-                            log.error(
-                                f'Failed to release lock for resource: resource_uri={resource[0]}, resource_type={resource[1]}'
-                            )
 
     @staticmethod
     def _get_share_data_and_items(session, share_uri, status, healthStatus=None):
@@ -422,28 +391,3 @@ class SharingService:
             session=session, share_uri=share_uri, status=[status], healthStatus=[healthStatus]
         )
         return share_data, share_items
-
-    @staticmethod
-    def acquire_lock_with_retry(
-        resources: List[Tuple[str, str]], session: Session, acquired_by_uri: str, acquired_by_type: str
-    ):
-        for attempt in range(MAX_RETRIES):
-            try:
-                log.info(f'Attempting to acquire lock for resources {resources} by share {acquired_by_uri}...')
-                lock_acquired = ResourceLockRepository.acquire_locks(
-                    resources, session, acquired_by_uri, acquired_by_type
-                )
-                if lock_acquired:
-                    return True
-
-                log.info(
-                    f'Lock for one or more resources {resources} already acquired. Retrying in {RETRY_INTERVAL} seconds...'
-                )
-                sleep(RETRY_INTERVAL)
-
-            except Exception as e:
-                log.error('Error occurred while retrying acquiring lock:', e)
-                return False
-
-        log.info(f'Max retries reached. Failed to acquire lock for one or more resources {resources}')
-        return False
